split src code
