+++ lot_list.php
@@ -1,5 +1,5 @@
 <?php
-require_once 'src/common.php';
+require_once 'src/include/common.php';
 
 require_once 'src/lot_queries.php';
 $lots = get_all_open_lots($connection, $cat_id);
