+++ src/utils/string.php
