+++ templates/page/register.php
@@ -1,6 +1,6 @@
 <?php
-require_once 'src/html_utils.php';
-require_once 'src/links.php';
+require_once 'src/utils/html.php';
+require_once 'src/utils/links.php';
 ?>
 <form class="form container <?= mark_if_true(!empty($errors), 'form--invalid') ?>"
       enctype="multipart/form-data" action="<?= get_register_page_link() ?>" method="post">
